+++ src/oci/regions.py
@@ -11,6 +11,7 @@ REGIONS_SHORT_NAMES = {
     'yyz': 'ca-toronto-1'
 }
 REGION_REALMS = {
+    'ap-seoul-1': 'oc1',
     'ap-tokyo-1': 'oc1',
     'us-phoenix-1': 'oc1',
     'us-ashburn-1': 'oc1',
@@ -31,6 +32,7 @@ REALMS = {
     'oc3': 'oraclegovcloud.com'
 }
 REGIONS = [
+    "ap-seoul-1",
     "ap-tokyo-1",
     "us-phoenix-1",
     "us-ashburn-1",